+++ tuxbox/neutrino/src/driver/rcinput.h
@@ -182,7 +182,7 @@ class CRCInput
 		long long calcTimeoutEnd_MS( int Timeout );
 
 		void getMsgAbsoluteTimeout(uint *msg, uint* data, unsigned long long *TimeoutEnd, bool bAllowRepeatLR= false);
-		void getMsg(uint *msg, uint* data, int Timeout, bool bAllowRepeatLR= false);     //get message, timeout in 1/10 secs :)
+		void getMsg(uint *msg, unsigned long long* data, int Timeout, bool bAllowRepeatLR= false);     //get message, timeout in 1/10 secs :)
 		void getMsg_ms(uint *msg, uint* data, int Timeout, bool bAllowRepeatLR= false);     //get message, timeout in msecs :)
 		void getMsg_us(uint *msg, uint* data, unsigned long long Timeout, bool bAllowRepeatLR= false);     //get message, timeout in µsecs :)
 		void postMsg(uint msg, uint data, bool Priority = true );  // push message back into buffer
