+++ tuxbox/neutrino/src/gui/movieplayer.cpp
@@ -215,7 +215,7 @@ int CMoviePlayerGui::show()
 	//TODO: Own LCD support
 	CLCD::getInstance()->setMode(CLCD::MODE_MP3);
 
-	uint msg; uint data;
+	uint msg; uint64_t data;
 
 	bool loop=true;
 	bool update=true;
